services/worker/requirements.txt
imapclient
psycopg[binary]
sqlalchemy
alembic
httpx
python-dotenv
beautifulsoup4
pgvector
